Add Swift Package Manager support to ConfigCat SDK for the Apple platforms (#11)
* Initial support for Swift Package Manager

* Consolidate all targets into two targets

* Restore ConfigCat Coverage scheme

* Update ConfigCat.podspec

Updated the source_files property in the podspec to match the location of the source code.

* Update Travis configuration

Updated the Travis configuration to use the shared ConfigCat scheme

* Remove processing of resources from Package.swift

* Change version to a constant

Changed the version to be a constant as infoDictionary?["CFBundleShortVersionString"] isn’t available when working with the Swift Package Manager as of Swift 5.3.  Additionally updated the deployment documentation.

* Run tests in parallel and random order

Improve the tests by running them in parallel and random order.

* Bump version

Bumped the version to 7.2.0 to match the description in the README.md.

* Update TravisCI macOS image

Updated the TravisCI macOS image to use Xcode 12.2. Updated all xcodebuild accordingly to support the change of Xcode.

* Fix cmdline compile failure

Fixed cmdline compilation due to missing dependecy from test target to framework target.